refactor to use Trainer api over Drivers
