+++ allennlp/experiments/__init__.py
@@ -1,3 +1,1 @@
-from allennlp.experiments.driver import Driver
-from allennlp.experiments.drivers.train_driver import TrainDriver
 from allennlp.experiments.registry import Registry
