+++ allennlp/training/trainer.py
@@ -1,16 +1,169 @@
+from typing import Optional
+import logging
+import os
+import shutil
+
+import torch
+import tqdm
+
 from allennlp.common import Params
-from allennlp.data import Dataset
+from allennlp.common.checks import ConfigurationError
+from allennlp.data.data_iterator import DataIterator
 from allennlp.training.model import Model
+from allennlp.data import Dataset
+from allennlp.common.tensor import arrays_to_variables
+
+logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
 
 class Trainer:
-    # Mark is working on actually implementing this - this is just a suggested API to make the
-    # driver simple.
-    def train_model(self,
-                    model: Model,
-                    train_dataset: Dataset,
-                    validation_dataset: Dataset):
-        pass
+    def __init__(self,
+                 model: Model,
+                 optimizer: torch.optim.Optimizer,
+                 iterator: DataIterator,
+                 train_dataset: Dataset,
+                 validation_dataset: Optional[Dataset] = None,
+                 patience: int = 2,
+                 batch_size: int = 32,
+                 num_epochs: int = 20,
+                 serialization_prefix: Optional[str] = None,
+                 cuda_device: int = -1) -> None:
+        """
+        Parameters
+        ----------
+        model : ``Model``, required.
+            An AllenNLP model to be optimized. Pytorch Modules can also be optimized if
+            their ``forward`` method returns a dictionary with a "loss" key, containing a
+            scalar tensor representing the loss function to be optimized.
+        optimizer : ``torch.nn.Optimizer``, required.
+            An instance of a Pytorch Optimizer, instantiated with the parameters of the
+            model to be optimized.
+        iterator : ``DataIterator``, required.
+            A method for iterating over a ``Dataset``, yielding padded indexed batches.
+        train_dataset : ``Dataset``, required.
+            A ``Dataset`` to train on. The dataset should have already been indexed.
+        validation_dataset : ``Dataset``, optional, (default = None).
+            A ``Dataset`` to evaluate on. The dataset should have already been indexed.
+        patience : int, optional (default=2)
+            Number of epochs to be patient before early stopping.  I.e., if the ``validation_metric``
+            does not improve for this many epochs, we will stop training.
+        batch_size : int, optional (default = 32)
+            Batch size to use when training.
+        num_epochs : int, optional (default = 20)
+            Number of training epochs.
+        serialization_prefix : str, optional (default=None)
+            Path to directory for saving and loading model files. Models will not be saved if
+            this parameter is not passed.
+        cuda_device : int, optional (default = -1)
+            An integer specifying the CUDA device to use. If -1, the CPU is used.
+            Multi-gpu training is not currently supported, but will be once the
+            Pytorch DataParallel API stabilises.
+        """
+        self._model = model
+        self._iterator = iterator
+        self._optimizer = optimizer
+        self._train_dataset = train_dataset
+        self._validation_dataset = validation_dataset
+
+        self._patience = patience
+        self._batch_size = batch_size
+        self._num_epochs = num_epochs
+        self._serialization_prefix = serialization_prefix
+        self._cuda_device = cuda_device
+
+        if self._cuda_device >= 0:
+            self._model = self._model.cuda(self._cuda_device)
+
+    def train(self) -> None:
+        epoch_counter = 0
+        # Resume from serialization path if it contains a saved model.
+        if self._serialization_prefix is not None:
+            if any(["model_state_epoch_" in x
+                    for x in os.listdir(self._serialization_prefix)]):
+                epoch_counter = self._restore_checkpoint()
+
+        for epoch in range(epoch_counter, self._num_epochs):
+            train_loss = 0.0
+            val_loss = 0.0
+
+            # Set the model to "train" mode.
+            self._model.train()
+            train_generator = self._iterator(self._train_dataset, num_epochs=1)
+            for batch in tqdm.tqdm(train_generator):
+                tensor_batch = arrays_to_variables(batch, self._cuda_device)
+                self._optimizer.zero_grad()
+                output_dict = self._model.forward(**tensor_batch)
+                try:
+                    loss = output_dict["loss"]
+                    loss.backward()
+                    train_loss += loss.data.numpy()
+                except KeyError:
+                    raise ConfigurationError("The model you are trying to optimize does not contain a"
+                                             " 'loss' key in the output of model.forward(inputs).")
+                self._optimizer.step()
+
+            if self._validation_dataset is not None:
+                # Switch to evaluation mode.
+                self._model.eval()
+                val_generator = self._iterator(self._validation_dataset, num_epochs=1)
+                for batch in tqdm.tqdm(val_generator):
+                    tensor_batch = arrays_to_variables(batch, self._cuda_device)
+                    val_output_dict = self._model.forward(tensor_batch)
+                    loss = val_output_dict["loss"]
+                    val_loss += loss.data.numpy()
+
+            # TODO(Mark): Add user specified metrics here, maybe a "metrics" key?
+            logger.info("Training Loss: %3f    Validation Loss: %3f ", train_loss, val_loss)
+            if self._serialization_prefix:
+                self._save_checkpoint(epoch)
+
+    def _save_checkpoint(self,
+                         epoch: int,
+                         is_best: Optional[bool] = None) -> None:
+        model_path = os.path.join(self._serialization_prefix, "model_state_epoch_{}.th".format(epoch))
+        model_state = self._model.state_dict()
+        torch.save(model_state, model_path)
+
+        training_state = {'epoch': epoch, 'optimizer': self._optimizer.state_dict()}
+        torch.save(training_state, os.path.join(self._serialization_prefix,
+                                                "training_state_epoch_{}.th".format(epoch)))
+        if is_best:
+            shutil.copy(model_path, os.path.join(model_path, "best.th"))
+
+    def _restore_checkpoint(self) -> int:
+
+        serialization_files = os.listdir(self._serialization_prefix)
+        model_checkpoints = [x for x in serialization_files if "model_state_epoch" in x]
+        epoch_to_load = max([int(x.split("model_state_epoch_")[-1].strip(".th")) for x in model_checkpoints])
+
+        model_state = torch.load(os.path.join(self._serialization_prefix,
+                                              "model_state_epoch_{}.th".format(epoch_to_load)))
+        training_state = torch.load(os.path.join(self._serialization_prefix,
+                                                 "training_state_epoch_{}.th".format(epoch_to_load)))
+        self._model.load_state_dict(model_state)
+        self._optimizer.load_state_dict(training_state["optimizer"])
+        return training_state["epoch"]
+
     @classmethod
-    def from_params(cls, params: Params) -> 'Trainer':
-        pass
+    def from_params(cls,
+                    model: Model,
+                    optimizer: torch.optim.Optimizer,
+                    iterator: DataIterator,
+                    train_dataset: Dataset,
+                    validation_dataset: Optional[Dataset],
+                    params: Optional[Params] = None) -> 'Trainer':
+
+        params = params or Params({})
+        patience = params.pop("patience", 2)
+        batch_size = params.pop("batch_size", 32)
+        num_epochs = params.pop("num_epochs", 20)
+        serialization_prefix = params.pop("serialization_prefix", None)
+        cuda_device = params.pop("cuda_device", -1)
+        params.assert_empty(cls.__name__)
+        return Trainer(model, optimizer, iterator,
+                       train_dataset, validation_dataset,
+                       patience=patience,
+                       batch_size=batch_size,
+                       num_epochs=num_epochs,
+                       serialization_prefix=serialization_prefix,
+                       cuda_device=cuda_device)
